Add figure reconstruction, plots for PC maps
If the 4D data is provided in an XYZxTime format, with no zeroes, that is the data was masked and then all the zeros removed the dramatically reduce memory usage:

data1_vec = reshape(data1, [],t_dim);
mask_ind = mask(:,:,:,1) ~= 0;

data1_vec_reduced = data1_vec(mask_ind,:);

and the user provides the original mask image in opt.reconmask, this will now reconstruct the 3d images from the reduced vector at the very end, thus generating all of the pretty plots for quite data evaluation. 

Also, this will now plot the timecourses of the PCs beneath the PC maps - though it is not as elegant as Kendrick's method.